Fix implicit conversion warning in OctoPrint.cpp

+++ src/slic3r/Utils/OctoPrint.cpp
@@ -141,7 +141,7 @@ std::string escape_string(const std::string& unescaped)
     std::string ret_val;
     CURL* curl = curl_easy_init();
     if (curl) {
-        char* decoded = curl_easy_escape(curl, unescaped.c_str(), unescaped.size());
+        char* decoded = curl_easy_escape(curl, unescaped.c_str(), static_cast<int>(unescaped.size()));
         if (decoded) {
             ret_val = std::string(decoded);
             curl_free(decoded);
